+++ src/backend/core/services/search/index.py
@@ -134,6 +134,7 @@ def index_message(message: models.Message) -> bool:
         "is_draft": message.is_draft,
         "is_trashed": message.is_trashed,
         "is_archived": message.is_archived,
+        "is_spam": message.is_spam,
         "is_starred": message.is_starred,
         "is_unread": message.is_unread,
         "is_sender": message.is_sender,