+++ src/backend/core/services/search/mapping.py
@@ -111,6 +111,7 @@ MESSAGE_MAPPING = {
             "is_draft": {"type": "boolean"},
             "is_trashed": {"type": "boolean"},
             "is_archived": {"type": "boolean"},
+            "is_spam": {"type": "boolean"},
             "is_starred": {"type": "boolean"},
             "is_unread": {"type": "boolean"},
             "is_sender": {"type": "boolean"},